style icons: tests
