+++ lib/assets/core/javascripts/cartodb3/components/form-components/editors/fill/input-color/assets-picker/assets-view.js
@@ -8,7 +8,7 @@ var PinIcons = require('../assets/pin-icons');
 var SimpleIcons = require('../assets/simple-icons');
 
 var UserAssetsView = require('./user-assets-tab');
-var OrganizationAssetsView = require('./organization-assets-tab');
+var OrganizationAssetsListView = require('./organization-assets-list-view');
 var OrganizationAssetsCollection = require('../../../../../../data/organization-assets-collection');
 var UploadAssetsView = require('./upload-assets-tab');
 
@@ -260,17 +260,21 @@ module.exports = CoreView.extend({
   },
 
   _createOrganizationUploadsView: function () {
-    this._organizationAssetsView = new OrganizationAssetsView({
-      model: this.model,
-      selectedAsset: this._selectedAsset,
-      title: _t('components.modals.add-asset.organization-uploads'),
-      organizationAssetCollection: this._organizationAssetCollection,
-      userAssetCollection: this._userAssetCollection
-    }).bind(this);
+    var view = new ScrollView({
+      createContentView: function () {
+        return new OrganizationAssetsListView({
+          title: _t('components.modals.add-asset.organization-uploads'),
+          model: this.model,
+          organizationAssetCollection: this._organizationAssetCollection,
+          userAssetCollection: this._userAssetCollection,
+          selectedAsset: this._selectedAsset,
+        });
+      }.bind(this)
+    });
 
     this._hideDisclaimer();
 
-    return this._organizationAssetsView;
+    return view;
   },
 
   _createUploadFileView: function () {
