+++ lib/assets/core/test/spec/cartodb3/components/form-components/editors/fill/input-color/asset-picker/organization-assets-list-view.spec.js
@@ -1,4 +1,5 @@
 var Backbone = require('backbone');
+var ConfigModel = require('../../../../../../../../../javascripts/cartodb3/data/config-model');
 var OrganizationAssetsCollection = require('../../../../../../../../../javascripts/cartodb3/data/organization-assets-collection');
 var OrganizationAssetsListView = require('../../../../../../../../../javascripts/cartodb3/components/form-components/editors/fill/input-color/assets-picker/organization-assets-list-view');
